+++ backend/spring-routie/src/main/java/routie/logging/config/LoggingConfig.java
@@ -10,7 +10,9 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import routie.logging.domain.LogDataBuilder;
 import routie.logging.infrastructure.aspect.RequestLoggingAspect;
 import routie.logging.infrastructure.interceptor.RequestLoggingInterceptor;
-import routie.logging.service.ClientRequestLogger;
+import routie.logging.infrastructure.strategy.DevLoggingStrategy;
+import routie.logging.infrastructure.strategy.ProdLoggingStrategy;
+import routie.logging.infrastructure.ClientRequestLogger;
 
 @Configuration
 @RequiredArgsConstructor
